concepts: refactor convertible_to out to is_convertible
This is the beginning of a relatively large concepts/type-traits refactor that is going to come in over the next while. First is factoring out convertible_to so it looks/works more like the standard library version so I can start converting detector into a more user-friendly version we can actually expose.  This is a bug for GEOSX and others, and should also make our type_traits more robust once it's all in.